backend/requirements.txt
django
django-environ
djangorestframework
django-cors-headers
django-filter
psycopg2-binary
Pillow
